End of Section # 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/Message';
 import { Pagination } from '../_models/Pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -18,7 +19,7 @@ export class MessagesComponent implements OnInit {
   loading = false;
 
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -35,9 +36,14 @@ export class MessagesComponent implements OnInit {
   
   deleteMessage(id: number) {
     //alert(id.toString());
-    this.messageService.deleteMessage(id).subscribe(() =>{
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('Delete message', 'Do you want to delete this message?','Cancel','Delete').subscribe(result => {
+      if (!result) {
+        this.messageService.deleteMessage(id).subscribe(() =>{
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
     })
+   
   }
 
   pageChanged(event: any) {
